Postgres log reports (#10)
* Install fluent to base image

* Update snapshot reference

* Enable automated logging of query plans

* Add fluentd.conf

* Setup fluentd to read postgres CSV logs

* Install pg fluent plugin

* Change postgres log file mode

* Implement postgres destination for fluentd

+++ deploy/impulse_pulumi/deploy_files/deploy_managed_pg.sh
@@ -2,14 +2,33 @@
 
 set -euxo pipefail
 
-sudo parted -s /dev/vdb mklabel gpt
-sudo parted -s /dev/vdb unit mib mkpart primary 0% 100%
-sudo mkfs.ext4 /dev/vdb1
-sudo mkdir /mnt/data
-echo | sudo tee -a /etc/fstab
-echo /dev/vdb1 /mnt/data ext4 defaults,noatime,nofail 0 0 | sudo tee -a /etc/fstab
-sudo mount /mnt/data
-sudo mkdir /mnt/data/postgres
-sudo chown postgres:postgres /mnt/data/postgres
+sudo ufw allow 5432
+# Explicitly gate the block storage setup for now to prevent accidental 
+# reconfiguration.
+if [[ ${SETUP_BLOCK_STORAGE} -eq 1 ]]; then
+  # Setup the block storage device as storage area for the "userdata" tablespace.
+  # postgresql_managed.conf already has "userdata" as the default tablespace.
+  sudo parted -s /dev/vdb mklabel gpt
+  sudo parted -s /dev/vdb unit mib mkpart primary 0% 100%
+  sudo mkfs.ext4 /dev/vdb1
+  sudo mkdir /mnt/data
+  echo | sudo tee -a /etc/fstab
+  echo /dev/vdb1 /mnt/data ext4 defaults,noatime,nofail 0 0 | sudo tee -a /etc/fstab
+  sudo mount /mnt/data
+  sudo mkdir /mnt/data/postgres
+  sudo chown postgres:postgres /mnt/data/postgres
+  sudo -u postgres psql -c "CREATE TABLESPACE userdata LOCATION '/mnt/data/postgres'"
+fi
 
-sudo -u postgres psql -c "CREATE TABLESPACE userdata LOCATION '/mnt/data/postgres'"
+sudo fluent-gem install fluent-plugin-postgresql-csvlog --no-document
+sudo fluent-gem install fluent-plugin-sql --no-document
+sudo fluent-gem install pg --no-document
+
+sudo sed --in-place "s/POSTGRES_HOST/${POSTGRES_HOST}/" /etc/fluent/fluentd.conf
+sudo sed --in-place "s/POSTGRES_PORT/${POSTGRES_PORT}/" /etc/fluent/fluentd.conf
+sudo sed --in-place "s/POSTGRES_DB/${POSTGRES_DB}/" /etc/fluent/fluentd.conf
+sudo sed --in-place "s/POSTGRES_USER/${POSTGRES_USER}/" /etc/fluent/fluentd.conf
+sudo sed --in-place "s/POSTGRES_PW/${POSTGRES_PW}/" /etc/fluent/fluentd.conf
+sudo sed --in-place "s/POSTGRES_TABLENAME/${POSTGRES_TABLENAME}/" /etc/fluent/fluentd.conf
+
+sudo systemctl restart postgresql fluentd
